Initial release!
This release provides an early release version of a library intended
for use with the processing of EZproxy related files and sessions.
This library was developed specifically to support the development of
an in-progress application, so the full context may not be entirely
clear until that application is released (currently pending review).

ADDED

- generate a list of audit records for session-related events
  - for all usernames
  - for a specific username

- generate a list of active sessions using audit log
  - using entires without a corresponding logout event type

- generate a list of active sessions using active file
  - for all usernames
  - for a specific username

- terminate user sessions
  - single user session
  - bulk user sessions

- Go modules support (vs classic `GOPATH` setup)

MISSING

- Anything to do with traffic log entries
- Examples