Doc comment: Fix name of MatchingUserSessions func
This was missed when I renamed the method from UserSessions
to MatchingUserSessions in an attempt to better clarify
the purpose of the method.

The doc comments for each implementation of the interface
method are correct, but I missed the change on the doc
comments for the interface method entry itself.